refactor(chain): share strict data snapshot boundary + prove barrel types
Address the open PR #1810 review threads on the strict finalized RPC transport.

r3610948521 (critical): the barrel type-export coverage lived only in a vitest
test, where type-only imports erase before runtime, so a dropped type re-export
still passed. Add packages/chain/tsconfig.type-tests.json + a `test:types` script
wired into `build` (tsc && pnpm run test:types), and a real compile-time proof
(test/rfc64-strict-rpc-public-api.typecheck.ts) that the strict adapter public
API — values AND types — is reachable through src/index.ts. Deleting a type
re-export now fails `test:types` (TS2305). The vitest test keeps the runtime
value assertions and no longer relies on erased type-only imports.

r3609986960 + r3610948522: the transport hand-rolled duplicate isPlainRecord,
dense-array/canonical-index, and exact-record validators, and shared the request
snapshot by exporting snapshotCurrentFinalizedEvmCallRequestV1 from the
router module. Extract one package-internal boundary
(current-finalized-strict-snapshot.ts) holding those primitives and the request
snapshotter; both current-finalized-evm-call.ts and the strict transport now
consume it. The router no longer exports the snapshotter and the barrel is
unchanged. Pure move: exact validation behavior and error messages preserved
(snapshotDenseDataArray carries each caller's domain message), with the full
chain unit suite (733) passing unedited.

r3610320608 (partial): config/data snapshotting is now its own focused unit and
the numbered-outcome classifier was extracted earlier; the remaining split
(JSON-RPC client, anchor/EVM decoders, block-reference execution, failover
orchestration) cascades into the fail-closed error taxonomy and reorg/failover
control flow and is left for a dedicated follow-up rather than a review-fix pass.

Preserves exact fail-closed snapshot semantics. Note: current-finalized-evm-call.ts
(PR #1808's file, already diverged on this branch) changed further here.
